pyproject.toml
[tool.poetry]
name = "dump_in"
version = "0.1.0"
description = ""
authors = ["SaJH <wogur981208@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
gunicorn = "^21.2.0"
requests = "^2.31.0"
cryptography = "^41.0.4"
drf-yasg = "^1.21.7"
djangorestframework-camel-case = "^1.4.2"
djangorestframework-simplejwt = "5.3.0"
pydantic = "^2.4.2"
django-filter = "^23.3"
psycopg2-binary = "^2.9.9"
django-prometheus = "^2.3.1"
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
django-redis = "^5.4.0"
slack-sdk = "^3.23.0"
aiohttp = "^3.8.6"
bcrypt = "^4.0.1"
oauthlib = "^3.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-django = "^4.7.0"
pytest-mock = "^3.12.0"
pytest-xdist = "^3.4.0"

factory-boy = "^3.3.0"
faker = "^19.6.1"

django-debug-toolbar = "^4.2.0"

mypy = "^1.5.1"

types-pillow = "^10.1.0.2"
django-stubs = "^4.2.4"
djangorestframework-stubs = "^3.14.2"
celery-stubs = "^0.1.3"
drf-yasg-stubs = "^0.1.4"

flake8 = "^6.1.0"
isort = "^5.12.0"
black = "^23.9.1"
pre-commit = "^3.4.0"
coverage = "^7.3.2"
pytest-cov = "^4.1.0"

[tool.black]
line-length = 140
target-version = ['py311']
force-exclude = 'migrations'

[tool.isort]
profile = "black"
filter_files = true
skip_glob = ["*/migrations/*", "config/*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
